Function in EpanetUtils has been modified
EPANET20012.dll added. This is a compile for my version of dll using the
code where I added a debug macros to print the parameters. It should be
the same that Epanet2_64bit.dll.

Modified method stringToByteBuffer to fix the error with string sended
to native using Buffers.

Added a method floatToByteBuffer to convert the float[] in a
FloatBuffer.