Added a dynamic GUI to configure problem
Added the gui
Added annotation used when the dynamic gui is builded.
Remove epanet.jna package (Is used the library EpanetToolkit supply by
teacher)
Add DefaultMethod attribute to operator.
Added method to validate the class with metadata of gui (ReflectionUtils) and to get the name of problem at the same class before mentioned.