Change the signature of getResult method in Algorithm
Change the signature of getResult method in Algorithm to list of
solution. It change is needed because to show the result is more complex
validate if the solution is a object or a list of objects.
Modified the diagram and report to add the change of signature of
method.
Fix some word in javadoc