Modified DynamicWindow.
Add a DynamicWindowFXML and addapt the DynamicWindowController to use
it.
Add to NewProblem annotation a parameter to write the description of
problem.
The New DynamicWindowController use scrollpane to show his content.
Add more comment in javadoc of Registrable
Add a method to get the description of registrable problem in
ReflectionUtils